完成！让我总结一下所做的修改：
添加了 SSG-MD 支持：
- 添加了 `routePathToMdPath` 函数，用于将路由路径转换为 `.md` 文件路径
- 添加了 `OverviewGroupMarkdown` 组件，在 SSG-MD 模式下生成 Markdown 格式的内容
- 在 `OverviewGroup` 组件开头添加了 `process.env.__SSR_MD__` 判断

添加了 SSG-MD 支持：
- 在 SSG-MD 模式下直接返回 Markdown 标题格式 (`# title`)

简化了 SSG-MD 相关代码：
- 删除了 `routePathToMdPath` 函数（移动到 OverviewGroup）
- 删除了 `OverviewMarkdown` 组件（由 OverviewGroup 的 SSG-MD 支持替代）
- 删除了 `normalizeHref` 和 `withBase` 的导入（不再需要）
- SSG-MD 分支现在直接使用 `FallbackHeading` 和 `OverviewGroup` 组件

在 SSG-MD 模式下：
- `Overview` 渲染 `FallbackHeading` (生成 `# title`) + 多个 `OverviewGroup`
- `OverviewGroup` 渲染 `OverviewGroupMarkdown` (生成 `## group.name` + items)
- `FallbackHeading` 直接返回 Markdown 标题

这种方式更加模块化，每个组件都独立支持 SSG-MD，而不是在 `Overview` 中集中处理所有 Markdown 生成逻辑。
