+++ packages/core/src/theme/components/Overview/index.tsx
@@ -6,7 +6,6 @@ import type {
 } from '@rspress/core';
 import {
   isEqualPath,
-  routePathToMdPath,
   useI18n,
   usePageData,
   useSidebar,
@@ -25,58 +24,6 @@ import {
 } from '../Sidebar/utils';
 import './index.scss';
 import { findItemByRoutePath } from './utils';
-
-function OverviewMarkdown({
-  title,
-  groups,
-}: {
-  title: string;
-  groups: Group[];
-}) {
-  const lines: string[] = [];
-
-  lines.push(`# ${title}`);
-  lines.push('');
-
-  for (const group of groups) {
-    if (group.name) {
-      lines.push(`## ${group.name}`);
-      lines.push('');
-    }
-
-    for (const item of group.items) {
-      const itemTitle = item.link
-        ? `[${item.text}](${routePathToMdPath(item.link)})`
-        : `**${item.text}**`;
-      lines.push(`### ${itemTitle}`);
-      lines.push('');
-
-      // Render headers as a list
-      if (item.headers && item.headers.length > 0) {
-        for (const header of item.headers) {
-          const headerLink = item.link
-            ? `[${header.text}](${routePathToMdPath(item.link)}#${header.id})`
-            : header.text;
-          lines.push(`- ${headerLink}`);
-        }
-        lines.push('');
-      }
-
-      // Render custom items as a list
-      if (item.items && item.items.length > 0) {
-        for (const subItem of item.items) {
-          const subItemLink = subItem.link
-            ? `[${subItem.text}](${routePathToMdPath(subItem.link)})`
-            : subItem.text;
-          lines.push(`- ${subItemLink}`);
-        }
-        lines.push('');
-      }
-    }
-  }
-
-  return <>{lines.join('\n')}</>;
-}
 
 // Utility function to normalize text for search
 const normalizeText = (s: string) => s.toLowerCase().replace(/-/g, ' ');
@@ -294,7 +241,14 @@ export function Overview(props: {
   const overviewTitle = title || 'Overview';
 
   if (process.env.__SSR_MD__) {
-    return <OverviewMarkdown title={overviewTitle} groups={groups} />;
+    return (
+      <>
+        <FallbackHeading level={1} title={overviewTitle} />
+        {groups.map(group => (
+          <OverviewGroup key={group?.name} group={group} />
+        ))}
+      </>
+    );
   }
 
   return (
