+++ packages/core/src/theme/components/OverviewGroup/index.tsx
@@ -1,9 +1,51 @@
 import { Link, renderInlineMarkdown } from '@theme';
 import './index.scss';
 import type { Header } from '@rspress/core';
+import { routePathToMdPath } from '@rspress/core/runtime';
 import { FallbackHeading, SvgWrapper } from '@theme';
 import { useMemo } from 'react';
 import IconPlugin from './icons/plugin.svg';
+
+function OverviewGroupMarkdown({ group }: { group: Group }) {
+  const lines: string[] = [];
+
+  if (group.name) {
+    lines.push(`## ${group.name}`);
+    lines.push('');
+  }
+
+  for (const item of group.items) {
+    const itemTitle = item.link
+      ? `[${item.text}](${routePathToMdPath(item.link)})`
+      : `**${item.text}**`;
+    lines.push(`### ${itemTitle}`);
+    lines.push('');
+
+    // Render headers as a list
+    if (item.headers && item.headers.length > 0) {
+      for (const header of item.headers) {
+        const headerLink = item.link
+          ? `[${header.text}](${routePathToMdPath(item.link)}#${header.id})`
+          : header.text;
+        lines.push(`- ${headerLink}`);
+      }
+      lines.push('');
+    }
+
+    // Render custom items as a list
+    if (item.items && item.items.length > 0) {
+      for (const subItem of item.items) {
+        const subItemLink = subItem.link
+          ? `[${subItem.text}](${routePathToMdPath(subItem.link)})`
+          : subItem.text;
+        lines.push(`- ${subItemLink}`);
+      }
+      lines.push('');
+    }
+  }
+
+  return <>{lines.join('\n')}</>;
+}
 
 export interface GroupItem {
   text: string;
@@ -20,6 +62,10 @@ export interface Group {
 }
 
 export const OverviewGroup = ({ group }: { group: Group }) => {
+  if (process.env.__SSR_MD__) {
+    return <OverviewGroupMarkdown group={group} />;
+  }
+
   const { itemsWithContent, itemsWithoutContent } = useMemo(() => {
     // Separate items into those with content and those without
     const itemsWithContent = group.items.filter(
